requirements.txt
Flask==2.3.3
Flask-MySQL==1.5.2
Flask-CORS==4.0.0
mysql-connector-python==8.1.0
python-dotenv==1.0.0